requirements.txt
flask_swagger_ui
azure-devops